Added missing exports from index.ts
I probably need to create a script to automatically create an updated index.ts
file at this point. Also, this update removes the useless memo() wrapper around
the `Option` component. Why is it useless? Because we use the React Compiler.
(and either way, it was just a fragment I forgot to remove during testing of
v0.4.0)